requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Flask==2.2.2
frozenlist==1.3.1
gevent==22.10.1
greenlet==1.1.3.post0
grequests==0.6.0
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.19.5
psycopg2==2.9.4
pycparser==2.21
pycryptodome==3.15.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pywin32==304
requests==2.28.1
rlp==2.0.1
six==1.16.0
tomli==2.0.1
toolz==0.12.0
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
Werkzeug==2.2.2
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.0
